Online Install bug fix

+++ RemoveDataFile.ps1
@@ -1,5 +1,10 @@
-﻿$ApDir = Join-Path $env:LOCALAPPDATA 'InterruptToDo'
+﻿# Remove data
 $ApFile = Join-Path $ApDir 'ToDoTasks.json'
+if( Test-Path $ApFile ){
+	Remove-Item $ApFile
+}
 
-Remove-Item $ApFile
-Remove-Item $ApDir
+$ApDir = Join-Path $env:LOCALAPPDATA 'InterruptToDo'
+if( Test-Path $ApDir ){
+	Remove-Item $ApDir
+}
